Make everything but futures an optional feature
