+++ .travis.yml
@@ -22,16 +22,25 @@ matrix:
   - env: TARGET=i686-unknown-freebsd
   - env: TARGET=i686-unknown-linux-gnu
 
-  # Test some combinations of enabled features.
+  # Test combinations of enabled features.
   - rust: stable
     script: |
-      cargo check --no-default-features --features threadpool
-  - rust: stable
-    script: |
-      cargo check --no-default-features --features threadpool,tcp
-  - rust: stable
-    script: |
-      cargo check --no-default-features --features threadpool,current-thread,udp,codec
+      shopt -s expand_aliases
+      alias check="cargo check --no-default-features"
+      check
+      check --features codec
+      check --features current-thread
+      check --features fs
+      check --features io
+      check --features reactor
+      check --features tcp
+      check --features threadpool
+      check --features timer
+      check --features udp
+      check --features uds
+      check --features current-thread,reactor,timer
+      check --features threadpool,reactor,timer
+      check --features current-thread,threadpool,reactor,timer
 
   # Test the async / await preview. We don't want to block PRs on this failing
   # though.
